requirements.txt
APScheduler==3.6.3
Django==2.2
django-apscheduler==0.4.2
djangorestframework==3.10.2
joblib==0.13.2
lightgbm==3.0.0
numpy==1.17.0
pandas==0.25.0
requests==2.24.0
scikit-learn==0.21.3